WCMS-17042: Dataset Tabs layout
